Silver-Tier/requirements.txt
watchdog==4.0.0
python-dotenv==1.0.0
google-auth==2.16.0
google-api-python-client==2.80.0
google-auth-oauthlib==1.0.0
mcp==1.0.0
schedule==1.2.0
anthropic>=0.40.0
requests>=2.32.0
playwright>=1.40.0
